requirements.txt
python-telegram-bot==21.2
requests==2.32.3
python-dotenv==1.0.0
pyTelegramBotAPI==4.18.1
